Split audio, UI, and drawing helpers into modules
